Refactor : Info items and edit fields are separate
